LayoutTests/fast/html/details-click-controls-expected.txt

PASS details.open is false
PASS details.open is false
PASS details.open is false
PASS details.open is false
PASS details.open is true
